+++ libraries/dfpUtils/dfpUtils.js
@@ -1,0 +1,13 @@
+/** Safe defaults which work on pretty much all video calls. */
+export const DEFAULT_DFP_PARAMS = {
+  env: 'vp',
+  gdfp_req: 1,
+  output: 'vast',
+  unviewed_position_start: 1,
+}
+
+export const DFP_ENDPOINT = {
+  protocol: 'https',
+  host: 'securepubads.g.doubleclick.net',
+  pathname: '/gampad/ads'
+}
